+++ nlcpy/ve_kernel/cblas_wrapper.c
@@ -3,10 +3,10 @@
 # * The source code in this file is developed independently by NEC Corporation.
 #
 # # NLCPy License #
-# 
+#
 #     Copyright (c) 2020-2021 NEC Corporation
 #     All rights reserved.
-#     
+#
 #     Redistribution and use in source and binary forms, with or without
 #     modification, are permitted provided that the following conditions are met:
 #     * Redistributions of source code must retain the above copyright notice,
@@ -17,7 +17,7 @@
 #     * Neither NEC Corporation nor the names of its contributors may be
 #       used to endorse or promote products derived from this software
 #       without specific prior written permission.
-#     
+#
 #     THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
 #     ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
 #     WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -31,9 +31,16 @@
 #
 */
 
+#include <stdio.h>
+#include <stdint.h>
+#include <stdbool.h>
+#include <stdlib.h>
+#include <limits.h>
+#include <alloca.h>
+#include <assert.h>
 
-#include <cblas.h>
 #include "nlcpy.h"
+#include <inc_i64/cblas.h>
 
 
 uint64_t wrapper_cblas_sdot(ve_arguments *args, int32_t *psw)
@@ -62,14 +69,14 @@ uint64_t wrapper_cblas_sdot(ve_arguments *args, int32_t *psw)
     }
     float *pz = (float *)z->ve_adr;
     if (pz == NULL) {
-       return (uint64_t)NLCPY_ERROR_MEMORY; 
+       return (uint64_t)NLCPY_ERROR_MEMORY;
     }
     assert(x->ndim <= 1);
     assert(y->ndim <= 1);
     assert(z->ndim <= 1);
     assert(x->size == y->size);
 
-    *pz = cblas_sdot(x->size, px, x->strides[0] / x->itemsize, 
+    *pz = cblas_sdot(x->size, px, x->strides[0] / x->itemsize,
                             py, y->strides[0] / y->itemsize);
 } /* omp single */
     retrieve_fpe_flags(psw);
@@ -102,14 +109,14 @@ uint64_t wrapper_cblas_ddot(ve_arguments *args, int32_t *psw)
     }
     double *pz = (double *)z->ve_adr;
     if (pz == NULL) {
-       return (uint64_t)NLCPY_ERROR_MEMORY; 
+       return (uint64_t)NLCPY_ERROR_MEMORY;
     }
     assert(x->ndim <= 1);
     assert(y->ndim <= 1);
     assert(z->ndim <= 1);
     assert(x->size == y->size);
 
-    *pz = cblas_ddot(x->size, px, x->strides[0] / x->itemsize, 
+    *pz = cblas_ddot(x->size, px, x->strides[0] / x->itemsize,
                             py, y->strides[0] / y->itemsize);
 } /* omp single */
     retrieve_fpe_flags(psw);
@@ -142,7 +149,7 @@ uint64_t wrapper_cblas_cdotu_sub(ve_arguments *args, int32_t *psw)
     }
     float _Complex *pz = (float _Complex *)z->ve_adr;
     if (pz == NULL) {
-       return (uint64_t)NLCPY_ERROR_MEMORY; 
+       return (uint64_t)NLCPY_ERROR_MEMORY;
     }
     assert(x->ndim <= 1);
     assert(y->ndim <= 1);
@@ -150,7 +157,7 @@ uint64_t wrapper_cblas_cdotu_sub(ve_arguments *args, int32_t *psw)
     assert(x->size == y->size);
 
 
-    cblas_cdotu_sub(x->size, px, x->strides[0] / x->itemsize, 
+    cblas_cdotu_sub(x->size, px, x->strides[0] / x->itemsize,
                         py, y->strides[0] / y->itemsize, pz);
 } /* omp single */
     retrieve_fpe_flags(psw);
@@ -183,7 +190,7 @@ uint64_t wrapper_cblas_zdotu_sub(ve_arguments *args, int32_t *psw)
     }
     double _Complex *pz = (double _Complex *)z->ve_adr;
     if (pz == NULL) {
-       return (uint64_t)NLCPY_ERROR_MEMORY; 
+       return (uint64_t)NLCPY_ERROR_MEMORY;
     }
     assert(x->ndim <= 1);
     assert(y->ndim <= 1);
@@ -191,7 +198,7 @@ uint64_t wrapper_cblas_zdotu_sub(ve_arguments *args, int32_t *psw)
     assert(x->size == y->size);
 
 
-    cblas_zdotu_sub(x->size, px, x->strides[0] / x->itemsize, 
+    cblas_zdotu_sub(x->size, px, x->strides[0] / x->itemsize,
                         py, y->strides[0] / y->itemsize, pz);
 } /* omp single */
     retrieve_fpe_flags(psw);
@@ -203,20 +210,20 @@ uint64_t wrapper_cblas_zdotu_sub(ve_arguments *args, int32_t *psw)
 
 uint64_t wrapper_cblas_sgemm(ve_arguments *args, int32_t *psw)
 {
-    const int32_t order = args->gemm.order;
-    const int32_t transA = args->gemm.transA;
-    const int32_t transB = args->gemm.transB;
-    const int32_t m = args->gemm.m;
-    const int32_t n = args->gemm.n;
-    const int32_t k = args->gemm.k;
+    const int64_t order = args->gemm.order;
+    const int64_t transA = args->gemm.transA;
+    const int64_t transB = args->gemm.transB;
+    const int64_t m = args->gemm.m;
+    const int64_t n = args->gemm.n;
+    const int64_t k = args->gemm.k;
     const float alpha = *((float *)nlcpy__get_scalar(&(args->gemm.alpha)));
     float* const a = (float *)args->gemm.a.ve_adr;
-    const int32_t lda = args->gemm.lda;
+    const int64_t lda = args->gemm.lda;
     float* const b = (float *)args->gemm.b.ve_adr;
-    const int32_t ldb = args->gemm.ldb;
+    const int64_t ldb = args->gemm.ldb;
     const float beta = *((float *)nlcpy__get_scalar(&(args->gemm.beta)));
     float* const c = (float *)args->gemm.c.ve_adr;
-    const int32_t ldc = args->gemm.ldc;
+    const int64_t ldc = args->gemm.ldc;
 
     if (a == NULL || b == NULL || c == NULL) {
         return NLCPY_ERROR_MEMORY;
@@ -224,25 +231,25 @@ uint64_t wrapper_cblas_sgemm(ve_arguments *args, int32_t *psw)
 
 
 #ifdef _OPENMP
-    const int32_t nt = omp_get_num_threads();
-    const int32_t it = omp_get_thread_num();
+    const int64_t nt = omp_get_num_threads();
+    const int64_t it = omp_get_thread_num();
 #else
-    const int32_t nt = 1;
-    const int32_t it = 0;
+    const int64_t nt = 1;
+    const int64_t it = 0;
 #endif /* _OPENMP */
 
-    const int32_t m_s = m * it / nt;
-    const int32_t m_e = m * (it + 1) / nt;
-    const int32_t m_d = m_e - m_s;
-    const int32_t n_s = n * it / nt;
-    const int32_t n_e = n * (it + 1) / nt;
-    const int32_t n_d = n_e - n_s;
-    
-    int32_t mode = 1;
-    if ( n > nt ) { 
+    const int64_t m_s = m * it / nt;
+    const int64_t m_e = m * (it + 1) / nt;
+    const int64_t m_d = m_e - m_s;
+    const int64_t n_s = n * it / nt;
+    const int64_t n_e = n * (it + 1) / nt;
+    const int64_t n_d = n_e - n_s;
+
+    int64_t mode = 1;
+    if ( n > nt ) {
         mode = 2;
     }
-    int32_t iar, iac, ibr, ibc, icr, icc;
+    int64_t iar, iac, ibr, ibc, icr, icc;
     if (transA == CblasNoTrans ) {
         iar = 1;
         iac = lda;
@@ -266,20 +273,20 @@ uint64_t wrapper_cblas_sgemm(ve_arguments *args, int32_t *psw)
     }
 
     if (order == CblasColMajor) {
-        if ( mode == 1 ) { 
+        if ( mode == 1 ) {
             // split 'm'
-            cblas_sgemm(order, transA, transB, m_d, n, k, alpha, a + m_s * iar, lda, b, ldb, beta, c + m_s * icr, ldc);  
+            cblas_sgemm(order, transA, transB, m_d, n, k, alpha, a + m_s * iar, lda, b, ldb, beta, c + m_s * icr, ldc);
         } else {
             // split 'n'
-            cblas_sgemm(order, transA, transB, m, n_d, k, alpha, a, lda, b + n_s * ibc, ldb, beta, c + n_s * icc, ldc);  
+            cblas_sgemm(order, transA, transB, m, n_d, k, alpha, a, lda, b + n_s * ibc, ldb, beta, c + n_s * icc, ldc);
         }
     } else {
-        if ( mode == 1 ) { 
+        if ( mode == 1 ) {
             // split 'm'
-            cblas_sgemm(order, transA, transB, m_d, n, k, alpha, a + m_s * iac, lda, b, ldb, beta, c + m_s * icr, ldc);  
+            cblas_sgemm(order, transA, transB, m_d, n, k, alpha, a + m_s * iac, lda, b, ldb, beta, c + m_s * icr, ldc);
         } else {
             // split 'n'
-            cblas_sgemm(order, transA, transB, m, n_d, k, alpha, a, lda, b + n_s * ibr, ldb, beta, c + n_s * icc, ldc);  
+            cblas_sgemm(order, transA, transB, m, n_d, k, alpha, a, lda, b + n_s * ibr, ldb, beta, c + n_s * icc, ldc);
         }
     }
 
@@ -289,20 +296,20 @@ uint64_t wrapper_cblas_sgemm(ve_arguments *args, int32_t *psw)
 
 uint64_t wrapper_cblas_dgemm(ve_arguments *args, int32_t *psw)
 {
-    const int32_t order = args->gemm.order;
-    const int32_t transA = args->gemm.transA;
-    const int32_t transB = args->gemm.transB;
-    const int32_t m = args->gemm.m;
-    const int32_t n = args->gemm.n;
-    const int32_t k = args->gemm.k;
+    const int64_t order = args->gemm.order;
+    const int64_t transA = args->gemm.transA;
+    const int64_t transB = args->gemm.transB;
+    const int64_t m = args->gemm.m;
+    const int64_t n = args->gemm.n;
+    const int64_t k = args->gemm.k;
     const double alpha = *((double *)nlcpy__get_scalar(&(args->gemm.alpha)));
     double* const a = (double *)args->gemm.a.ve_adr;
-    const int32_t lda = args->gemm.lda;
+    const int64_t lda = args->gemm.lda;
     double* const b = (double *)args->gemm.b.ve_adr;
-    const int32_t ldb = args->gemm.ldb;
+    const int64_t ldb = args->gemm.ldb;
     const double beta = *((double *)nlcpy__get_scalar(&(args->gemm.beta)));
     double* const c = (double *)args->gemm.c.ve_adr;
-    const int32_t ldc = args->gemm.ldc;
+    const int64_t ldc = args->gemm.ldc;
 
     if (a == NULL || b == NULL || c == NULL) {
         return NLCPY_ERROR_MEMORY;
@@ -310,25 +317,25 @@ uint64_t wrapper_cblas_dgemm(ve_arguments *args, int32_t *psw)
 
 
 #ifdef _OPENMP
-    const int32_t nt = omp_get_num_threads();
-    const int32_t it = omp_get_thread_num();
+    const int64_t nt = omp_get_num_threads();
+    const int64_t it = omp_get_thread_num();
 #else
-    const int32_t nt = 1;
-    const int32_t it = 0;
+    const int64_t nt = 1;
+    const int64_t it = 0;
 #endif /* _OPENMP */
 
-    const int32_t m_s = m * it / nt;
-    const int32_t m_e = m * (it + 1) / nt;
-    const int32_t m_d = m_e - m_s;
-    const int32_t n_s = n * it / nt;
-    const int32_t n_e = n * (it + 1) / nt;
-    const int32_t n_d = n_e - n_s;
-    
-    int32_t mode = 1;
-    if ( n > nt ) { 
+    const int64_t m_s = m * it / nt;
+    const int64_t m_e = m * (it + 1) / nt;
+    const int64_t m_d = m_e - m_s;
+    const int64_t n_s = n * it / nt;
+    const int64_t n_e = n * (it + 1) / nt;
+    const int64_t n_d = n_e - n_s;
+
+    int64_t mode = 1;
+    if ( n > nt ) {
         mode = 2;
     }
-    int32_t iar, iac, ibr, ibc, icr, icc;
+    int64_t iar, iac, ibr, ibc, icr, icc;
     if (transA == CblasNoTrans ) {
         iar = 1;
         iac = lda;
@@ -352,20 +359,20 @@ uint64_t wrapper_cblas_dgemm(ve_arguments *args, int32_t *psw)
     }
 
     if (order == CblasColMajor) {
-        if ( mode == 1 ) { 
+        if ( mode == 1 ) {
             // split 'm'
-            cblas_dgemm(order, transA, transB, m_d, n, k, alpha, a + m_s * iar, lda, b, ldb, beta, c + m_s * icr, ldc);  
+            cblas_dgemm(order, transA, transB, m_d, n, k, alpha, a + m_s * iar, lda, b, ldb, beta, c + m_s * icr, ldc);
         } else {
             // split 'n'
-            cblas_dgemm(order, transA, transB, m, n_d, k, alpha, a, lda, b + n_s * ibc, ldb, beta, c + n_s * icc, ldc);  
+            cblas_dgemm(order, transA, transB, m, n_d, k, alpha, a, lda, b + n_s * ibc, ldb, beta, c + n_s * icc, ldc);
         }
     } else {
-        if ( mode == 1 ) { 
+        if ( mode == 1 ) {
             // split 'm'
-            cblas_dgemm(order, transA, transB, m_d, n, k, alpha, a + m_s * iac, lda, b, ldb, beta, c + m_s * icr, ldc);  
+            cblas_dgemm(order, transA, transB, m_d, n, k, alpha, a + m_s * iac, lda, b, ldb, beta, c + m_s * icr, ldc);
         } else {
             // split 'n'
-            cblas_dgemm(order, transA, transB, m, n_d, k, alpha, a, lda, b + n_s * ibr, ldb, beta, c + n_s * icc, ldc);  
+            cblas_dgemm(order, transA, transB, m, n_d, k, alpha, a, lda, b + n_s * ibr, ldb, beta, c + n_s * icc, ldc);
         }
     }
 
@@ -375,22 +382,22 @@ uint64_t wrapper_cblas_dgemm(ve_arguments *args, int32_t *psw)
 
 uint64_t wrapper_cblas_cgemm(ve_arguments *args, int32_t *psw)
 {
-    const int32_t order = args->gemm.order;
-    const int32_t transA = args->gemm.transA;
-    const int32_t transB = args->gemm.transB;
-    const int32_t m = args->gemm.m;
-    const int32_t n = args->gemm.n;
-    const int32_t k = args->gemm.k;
+    const int64_t order = args->gemm.order;
+    const int64_t transA = args->gemm.transA;
+    const int64_t transB = args->gemm.transB;
+    const int64_t m = args->gemm.m;
+    const int64_t n = args->gemm.n;
+    const int64_t k = args->gemm.k;
     const void *alpha = (void *)nlcpy__get_scalar(&(args->gemm.alpha));
     if (alpha == NULL) return (uint64_t)NLCPY_ERROR_MEMORY;
     float  _Complex* const a = (float  _Complex *)args->gemm.a.ve_adr;
-    const int32_t lda = args->gemm.lda;
+    const int64_t lda = args->gemm.lda;
     float  _Complex* const b = (float  _Complex *)args->gemm.b.ve_adr;
-    const int32_t ldb = args->gemm.ldb;
+    const int64_t ldb = args->gemm.ldb;
     const void *beta = (void *)nlcpy__get_scalar(&(args->gemm.beta));
     if (beta == NULL) return (uint64_t)NLCPY_ERROR_MEMORY;
     float  _Complex* const c = (float  _Complex *)args->gemm.c.ve_adr;
-    const int32_t ldc = args->gemm.ldc;
+    const int64_t ldc = args->gemm.ldc;
 
     if (a == NULL || b == NULL || c == NULL) {
         return NLCPY_ERROR_MEMORY;
@@ -398,25 +405,25 @@ uint64_t wrapper_cblas_cgemm(ve_arguments *args, int32_t *psw)
 
 
 #ifdef _OPENMP
-    const int32_t nt = omp_get_num_threads();
-    const int32_t it = omp_get_thread_num();
+    const int64_t nt = omp_get_num_threads();
+    const int64_t it = omp_get_thread_num();
 #else
-    const int32_t nt = 1;
-    const int32_t it = 0;
+    const int64_t nt = 1;
+    const int64_t it = 0;
 #endif /* _OPENMP */
 
-    const int32_t m_s = m * it / nt;
-    const int32_t m_e = m * (it + 1) / nt;
-    const int32_t m_d = m_e - m_s;
-    const int32_t n_s = n * it / nt;
-    const int32_t n_e = n * (it + 1) / nt;
-    const int32_t n_d = n_e - n_s;
-    
-    int32_t mode = 1;
-    if ( n > nt ) { 
+    const int64_t m_s = m * it / nt;
+    const int64_t m_e = m * (it + 1) / nt;
+    const int64_t m_d = m_e - m_s;
+    const int64_t n_s = n * it / nt;
+    const int64_t n_e = n * (it + 1) / nt;
+    const int64_t n_d = n_e - n_s;
+
+    int64_t mode = 1;
+    if ( n > nt ) {
         mode = 2;
     }
-    int32_t iar, iac, ibr, ibc, icr, icc;
+    int64_t iar, iac, ibr, ibc, icr, icc;
     if (transA == CblasNoTrans ) {
         iar = 1;
         iac = lda;
@@ -440,20 +447,20 @@ uint64_t wrapper_cblas_cgemm(ve_arguments *args, int32_t *psw)
     }
 
     if (order == CblasColMajor) {
-        if ( mode == 1 ) { 
+        if ( mode == 1 ) {
             // split 'm'
-            cblas_cgemm(order, transA, transB, m_d, n, k, alpha, a + m_s * iar, lda, b, ldb, beta, c + m_s * icr, ldc);  
+            cblas_cgemm(order, transA, transB, m_d, n, k, alpha, a + m_s * iar, lda, b, ldb, beta, c + m_s * icr, ldc);
         } else {
             // split 'n'
-            cblas_cgemm(order, transA, transB, m, n_d, k, alpha, a, lda, b + n_s * ibc, ldb, beta, c + n_s * icc, ldc);  
+            cblas_cgemm(order, transA, transB, m, n_d, k, alpha, a, lda, b + n_s * ibc, ldb, beta, c + n_s * icc, ldc);
         }
     } else {
-        if ( mode == 1 ) { 
+        if ( mode == 1 ) {
             // split 'm'
-            cblas_cgemm(order, transA, transB, m_d, n, k, alpha, a + m_s * iac, lda, b, ldb, beta, c + m_s * icr, ldc);  
+            cblas_cgemm(order, transA, transB, m_d, n, k, alpha, a + m_s * iac, lda, b, ldb, beta, c + m_s * icr, ldc);
         } else {
             // split 'n'
-            cblas_cgemm(order, transA, transB, m, n_d, k, alpha, a, lda, b + n_s * ibr, ldb, beta, c + n_s * icc, ldc);  
+            cblas_cgemm(order, transA, transB, m, n_d, k, alpha, a, lda, b + n_s * ibr, ldb, beta, c + n_s * icc, ldc);
         }
     }
 
@@ -463,22 +470,22 @@ uint64_t wrapper_cblas_cgemm(ve_arguments *args, int32_t *psw)
 
 uint64_t wrapper_cblas_zgemm(ve_arguments *args, int32_t *psw)
 {
-    const int32_t order = args->gemm.order;
-    const int32_t transA = args->gemm.transA;
-    const int32_t transB = args->gemm.transB;
-    const int32_t m = args->gemm.m;
-    const int32_t n = args->gemm.n;
-    const int32_t k = args->gemm.k;
+    const int64_t order = args->gemm.order;
+    const int64_t transA = args->gemm.transA;
+    const int64_t transB = args->gemm.transB;
+    const int64_t m = args->gemm.m;
+    const int64_t n = args->gemm.n;
+    const int64_t k = args->gemm.k;
     const void *alpha = (void *)nlcpy__get_scalar(&(args->gemm.alpha));
     if (alpha == NULL) return (uint64_t)NLCPY_ERROR_MEMORY;
     double _Complex* const a = (double _Complex *)args->gemm.a.ve_adr;
-    const int32_t lda = args->gemm.lda;
+    const int64_t lda = args->gemm.lda;
     double _Complex* const b = (double _Complex *)args->gemm.b.ve_adr;
-    const int32_t ldb = args->gemm.ldb;
+    const int64_t ldb = args->gemm.ldb;
     const void *beta = (void *)nlcpy__get_scalar(&(args->gemm.beta));
     if (beta == NULL) return (uint64_t)NLCPY_ERROR_MEMORY;
     double _Complex* const c = (double _Complex *)args->gemm.c.ve_adr;
-    const int32_t ldc = args->gemm.ldc;
+    const int64_t ldc = args->gemm.ldc;
 
     if (a == NULL || b == NULL || c == NULL) {
         return NLCPY_ERROR_MEMORY;
@@ -486,25 +493,25 @@ uint64_t wrapper_cblas_zgemm(ve_arguments *args, int32_t *psw)
 
 
 #ifdef _OPENMP
-    const int32_t nt = omp_get_num_threads();
-    const int32_t it = omp_get_thread_num();
+    const int64_t nt = omp_get_num_threads();
+    const int64_t it = omp_get_thread_num();
 #else
-    const int32_t nt = 1;
-    const int32_t it = 0;
+    const int64_t nt = 1;
+    const int64_t it = 0;
 #endif /* _OPENMP */
 
-    const int32_t m_s = m * it / nt;
-    const int32_t m_e = m * (it + 1) / nt;
-    const int32_t m_d = m_e - m_s;
-    const int32_t n_s = n * it / nt;
-    const int32_t n_e = n * (it + 1) / nt;
-    const int32_t n_d = n_e - n_s;
-    
-    int32_t mode = 1;
-    if ( n > nt ) { 
+    const int64_t m_s = m * it / nt;
+    const int64_t m_e = m * (it + 1) / nt;
+    const int64_t m_d = m_e - m_s;
+    const int64_t n_s = n * it / nt;
+    const int64_t n_e = n * (it + 1) / nt;
+    const int64_t n_d = n_e - n_s;
+
+    int64_t mode = 1;
+    if ( n > nt ) {
         mode = 2;
     }
-    int32_t iar, iac, ibr, ibc, icr, icc;
+    int64_t iar, iac, ibr, ibc, icr, icc;
     if (transA == CblasNoTrans ) {
         iar = 1;
         iac = lda;
@@ -528,20 +535,20 @@ uint64_t wrapper_cblas_zgemm(ve_arguments *args, int32_t *psw)
     }
 
     if (order == CblasColMajor) {
-        if ( mode == 1 ) { 
+        if ( mode == 1 ) {
             // split 'm'
-            cblas_zgemm(order, transA, transB, m_d, n, k, alpha, a + m_s * iar, lda, b, ldb, beta, c + m_s * icr, ldc);  
+            cblas_zgemm(order, transA, transB, m_d, n, k, alpha, a + m_s * iar, lda, b, ldb, beta, c + m_s * icr, ldc);
         } else {
             // split 'n'
-            cblas_zgemm(order, transA, transB, m, n_d, k, alpha, a, lda, b + n_s * ibc, ldb, beta, c + n_s * icc, ldc);  
+            cblas_zgemm(order, transA, transB, m, n_d, k, alpha, a, lda, b + n_s * ibc, ldb, beta, c + n_s * icc, ldc);
         }
     } else {
-        if ( mode == 1 ) { 
+        if ( mode == 1 ) {
             // split 'm'
-            cblas_zgemm(order, transA, transB, m_d, n, k, alpha, a + m_s * iac, lda, b, ldb, beta, c + m_s * icr, ldc);  
+            cblas_zgemm(order, transA, transB, m_d, n, k, alpha, a + m_s * iac, lda, b, ldb, beta, c + m_s * icr, ldc);
         } else {
             // split 'n'
-            cblas_zgemm(order, transA, transB, m, n_d, k, alpha, a, lda, b + n_s * ibr, ldb, beta, c + n_s * icc, ldc);  
+            cblas_zgemm(order, transA, transB, m, n_d, k, alpha, a, lda, b + n_s * ibr, ldb, beta, c + n_s * icc, ldc);
         }
     }
